Fixed settings issue
